third_party/blink/web_tests/external/wpt/mediacapture-streams/MediaDevices-enumerateDevices-returned-objects.https-expected.txt

This is a testharness.js-based test.
FAIL enumerateDevices exposes mostly empty objects ahead of successful getUserMedia call assert_equals: deviceId is empty before capture expected "" but got "default"
PASS enumerateDevices exposes expected objects after successful getUserMedia call
Harness: the test ran to completion.
